Use socket rather than address to determine connection type
In sendmsg() and sendto() we were inspecting the sock_addr_t
structure to determine if the connection was Internet or not.

Since msg->msg_name is an optional value in sendmsg() and
sendto() this could result in crashes because we weren't ensuring
it was non-null.

Since it's optional we should have been inspecting the SO_DOMAIN
of the connection's socket anyway - it will always be there.

Part of the fix for:
http://code.google.com/p/torsocks/issues/detail?id=15