Issue 29: tordns no worky?
This was a regression caused by splitting tsocks.c into socks.c and
torsocks.c. The 'pool' variable was declared static in the socks.h
header file with the result that the value was no longer shared between
source files.

Define pool in socks.c and declare it extern in socks.h.

Based on patch from foobi..@gmail.com