Issue 27: Fix build system so 'make dist' works
"The current build system fails when trying to create a tarball using 'make dist'.
There are two problems here:

When creating a tarball for distribution with 'make dist', the header files at
src/*.h are not included and so the build fails.  These should be listed in
libtorsocks_la_SOURCES.

Similarly, not having the prefix dist_ for the installation directories in doc/
means that these will not be included in the distribution tarball.  Also, there
is no need to define custom *manpagedirs as the build system is man page aware."

Patch by Anthony Basile <blueness@gentoo.org>