Try both symbol names before showing library error
The correct library we need to find may be the symbol sym or __sym or
something else entirely that we can't handle yet. We want to try to find
the library using both naming conventions, because they are the most
common, before we display a (very repetitive) warning.

Adjust macro name, check if dlerror returns null. strdup dlerror after sym
not found.

Give the macro a more descriptive name and add it to torsocks namespace.
Also use a default "Not Found" reason if both dysym and dlerror return
null.

Factor out common macro