Multiple changes accross the code base
This is a pretty big commit because a lot of inter connected things had
to be changed across the code base.

The config file part is completely reengineered because it turns out the
torsocks.conf was not really "Tor centric" and was adding features that
are IMHO out of the scope of torsocks. For now only two options are
supported which are "TorAddress" and "TorPort" having a format like Tor
config file has.

The global configuration object is added also which is initialized
within the constructor of the library using the new config file API.
Some utils functions were added to support that as well.

Also, the connection.c/.h is improved.

At this commit, torsocks can parse a config file or use default values
if none found, do logging and connect to Tor using only SOCKS5.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>