Lookup symbols in libc in constructor
This is to make sure that before main() we get the libc pointer from the
system library.

This commit also adds mmap/munmap support for syscall(). A comment in
lib/syscall.c explains for what specific case this is needed.

Finally, stop looking up symbols that were already looked up in the
constructor.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>