Fix: bad include and duplicate syscall values
First, for the unistd.h include, remove the i386 and s390/x defined to
include asm-generic/unistd.h since this header is x64 specific! If some
syscall's values are missing, they will be nullified and handled
accordingly.

The second issue is that we set different bad values to an unknown
syscall value so we don't end up with duplicates in the torsocks syscall
wrapper switch case.

Fixes #18

Signed-off-by: David Goulet <dgoulet@ev0ke.net>