Fix: remove the use of IPv4 sockaddr in connect
This modifies the onion find by ip to use generic sockaddr structure.
Also, a utils call is added to extract the port from a sockaddr
structure depending on the family.

This is a step in the direction of fully supported IPv6 thus removing
the use of a cast sockaddr_in.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>