Add AllowOutboundLocalhost.
This addresses the tails problem with #8137 in that, with the options
set, all connections to the loopback address is allowed. Since this is a
massive foot + gun option for the users, it's off by default.

Note that to correctly fix MAPADDRESS and all the things that the Whonix
people are complaining about would require bringing back the old tsocks
"local" configuration option.

Signed-off-by: Yawning Angel <yawning@schwanenlied.me>
Signed-off-by: David Goulet <dgoulet@ev0ke.net>