Use the correct libc symbol for socket, __socket30, on newer NetBSD.
Otherwise, we will define the ELF symbol __socket30 due to the declaration of
socket in <sys/socket.h>, but try using dlsym to call the libc symbol
"socket", which in NetBSD is defined as a compatibility wrapper that calls
__socket30...which, because of ELF symbol resolution rules, will go back and
call our __socket30, leading to infinite recursion.

Instead, call the libc symbol "__socket30" explicitly so we get the real libc
system call stub instead of the compatibility wrapper.

(I don't know any way to automatically deduce the correct ELF symbol from the
toolchain, even though it is written in <sys/socket.h>.)

Signed-off-by: David Goulet <dgoulet@ev0ke.net>