Hijack execve() syscall to check for cap/setuid
If a binary is set with capabilities or is setuid/gid, the kernel will strip
out the LD_PRELOAD thus making torsocks useless. This is only working on Linux
at this point.

This is not a complete fix but it's a good start and useful to have. See
ticket #14322 for more information.

Patch from cypherpunks with minor syntax edit.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>