Fix memcpy() buffer overrun in gethostbyname()
An extra 12 bytes of uninitialized data from the stack was copied in the
static tsocks_he_addr object which is then returned in the hostent object by
gethostbyname().

First of all, this patch sets the right length to tsocks_he_addr which is 4
bytes that is the IPv4 address in network byte order.

Second, gethostbyname can take a valid IPv4 so check that before trying to
resolve so to speed up the call.

Thanks to guido for reporting this through our Hackerone bug bounty program.

Reported-by: Guido Vranken <guidovranken@gmail.com>
Signed-off-by: David Goulet <dgoulet@ev0ke.net>