Ignore stderr for getcap command
On a filesystem not supporting capabilities, an error is thrown out on
stderr which is undesirable and confusing for the user.

This commit throws stderr to /dev/null. The point of this getcap is to
actually see if any capabilities are enabled or none. In the context of
an error, there are none.

Fixes #21088

Signed-off-by: David Goulet <dgoulet@ev0ke.net>