socks5: Always use ATYP 0x03 for CONNECT command
Because of the SafeSocks parameter of Tor, if set, we can't pass a raw IP
address to Tor since it will believe we did a DNS resolution from the
application.

Now, thanks to #22461, Tor safely accepts an IPv4/IPv6 address withing a FQDN
connect request which avoids the SafeSocks warnings.

The #22461 wasn't backported which means that torsocks working with SafeSocks
is only possible in tor >= 0.3.2 stable series.

Fixes #23667

Signed-off-by: David Goulet <dgoulet@torproject.org>